SEA-specific build config
