+++ packages/cli/package.json
@@ -7,6 +7,7 @@
   },
   "files": [
     "dist",
+    "!dist/sea",
     "README.md",
     "LICENSE"
   ],
@@ -22,6 +23,7 @@
   },
   "scripts": {
     "build": "tsup",
+    "build:sea": "tsup --config tsup.sea.config.ts",
     "prepack": "cp ../../README.md ../../LICENSE .",
     "typecheck": "tsc",
     "test": "vitest run",
